# schema/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/users/users/{user_id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get User",
                "description": ":param request: The current request being processed.\n:param user_id: The unique identifier of the user to be retrieved.\n:param user_payload: The payload containing information about the user, typically extracted from the JWT token.\n:return: A UserResponse object containing the user's details if found, otherwise raises an HTTPException if the user is not found.",
                "operationId": "get_user_users_users__user_id__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/my_profile/": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "My Profile",
                "operationId": "my_profile_users_my_profile__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/users/modify_my_profile/": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Modify Profile",
                "operationId": "modify_profile_users_modify_my_profile__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModifyProfileRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModifyProfileResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/users/change_password/": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Change Password",
                "description": "Change the password of the authenticated user.\n\n:param request: The current request being processed.\n:param profile_request: A ChangePasswordRequest object containing the new password.\n:return: A ChangePasswordResponse indicating that the profile was updated successfully.\n:raises HTTPException: If the user does not exist or there is an error during password change.",
                "operationId": "change_password_users_change_password__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChangePasswordResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/users/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get Users",
                "operationId": "get_users_users_users_get",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Ids"
                        }
                    },
                    {
                        "name": "location_filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Location Filter"
                        }
                    },
                    {
                        "name": "name_filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Name Filter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUsersResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChangePasswordRequest": {
                "properties": {
                    "old_password": {
                        "type": "string",
                        "title": "Old Password"
                    },
                    "new_password": {
                        "type": "string",
                        "title": "New Password"
                    }
                },
                "type": "object",
                "required": [
                    "old_password",
                    "new_password"
                ],
                "title": "ChangePasswordRequest"
            },
            "ChangePasswordResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Profile updated successfully"
                    }
                },
                "type": "object",
                "title": "ChangePasswordResponse"
            },
...
            "Gender": {
                "type": "string",
                "enum": [
                    "male",
                    "female",
                    "other"
                ],
                "title": "Gender"
            },
...
            "GetUsersResponse": {
                "properties": {
                    "users_info": {
                        "items": {
                            "$ref": "#/components/schemas/UserShortInfo"
                        },
                        "type": "array",
                        "title": "Users Info"
                    }
                },
                "type": "object",
                "required": [
                    "users_info"
                ],
                "title": "GetUsersResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ModifyProfileRequest": {
                "properties": {
                    "fullName": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fullname"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "login": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Login"
                    },
                    "permanent_address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Permanent Address"
                    },
                    "current_address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Current Address"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Gender"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "phone_number": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phone Number"
                    },
                    "birthday": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Birthday"
                    },
                    "bio_header": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bio Header"
                    },
                    "bio_description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bio Description"
                    },
                    "interests": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Interests"
                    }
                },
                "type": "object",
                "title": "ModifyProfileRequest"
            },
            "ModifyProfileResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Profile updated successfully"
                    }
                },
                "type": "object",
                "title": "ModifyProfileResponse"
            },
...
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "fullName": {
                        "type": "string",
                        "title": "Fullname"
                    },
                    "permanent_address": {
                        "type": "string",
                        "title": "Permanent Address"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "member_since": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Member Since"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "phone_number": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Phone Number"
                    },
                    "current_address": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Current Address"
                    },
                    "birthday": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Birthday"
                    },
                    "bio_header": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bio Header"
                    },
                    "bio_description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bio Description"
                    },
                    "interests": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Interests"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "fullName",
                    "permanent_address",
                    "email",
                    "login",
                    "member_since",
                    "is_active",
                    "interests"
                ],
                "title": "UserResponse"
            },
            "UserShortInfo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "image_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Id"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "UserShortInfo"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "description": "Bearer token",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}